Rename IKVStorage.h -> IKeyValueStorage.h
